requirements.txt
# API Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ORM & DB drivers
sqlalchemy==2.0.30
aiomysql==0.2.0
asyncpg==0.29.0
alembic==1.13.1

# Settings
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# HTTP client
httpx==0.27.0

# Groq cloud inference
groq==0.9.0

# OpenAI SDK (used for NVIDIA NIM, Gemini, Cerebras OpenAI-compatible APIs)
openai==1.30.0

# Text-to-Speech (Microsoft Edge TTS, free, no key, includes Ghana voices)
edge-tts==6.1.10

# Security
argon2-cffi==23.1.0

# Rate limiting
slowapi==0.1.9

# Language detection
langdetect==1.0.9

# Input sanitization
bleach==6.1.0

# PDF generation
weasyprint==68.1

# Templates
jinja2==3.1.4

# Async file I/O (StaticFiles)
aiofiles==23.2.1

# Visual Search
chromadb>=0.6.0
sentence-transformers==3.0.0
Pillow==10.3.0
numpy<2.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
aiosqlite==0.20.0
